+++ app/src/Menu.tsx
@@ -2,52 +2,16 @@
 SPDX-License-Identifier: GPL-3.0-only */
 
 import { Fragment } from "react";
-import IconSVG from "./svg/icon.svg?react";
-import { useGlitch } from "react-powerglitch";
 import { Separator } from "@packages/cloud-react/lib/base/structure/Separator";
 import { Link, useLocation } from "react-router-dom";
 import { nameFromRoute, routeCategories } from "./config/routes";
-import { useTheme } from "./contexts/Theme";
+import "./styles/menu.scss";
 
 export const Menu = () => {
   const { pathname } = useLocation();
-  const { mode, toggleMode } = useTheme();
-
-  const glitch = useGlitch({
-    timing: {
-      duration: 5000,
-    },
-    glitchTimeSpan: {
-      start: 0.5,
-      end: 0.65,
-    },
-    shake: {
-      velocity: 2,
-      amplitudeX: 0.2,
-      amplitudeY: 0.2,
-    },
-  });
 
   return (
     <div className="menu">
-      <div className="icon" ref={glitch.ref}>
-        <IconSVG />
-      </div>
-      <section>
-        <button
-          className={`link${mode === "light" ? " selected" : ``}`}
-          onClick={() => toggleMode("light")}
-        >
-          Light
-        </button>
-        <button
-          className={`link${mode === "dark" ? " selected" : ``}`}
-          onClick={() => toggleMode("dark")}
-        >
-          Dark
-        </button>
-        <Separator />
-      </section>
       <section>
         {routeCategories.map(({ name, ...rest }, i) => (
           <Fragment key={`nav_${i}`}>
