+++ packages/cloud-core/lib/scss/base/modal/ModalCard/index.scss
@@ -28,10 +28,10 @@ SPDX-License-Identifier: GPL-3.0-only */
   }
 
   .header {
-    width: 100%;
     display: flex;
     align-items: center;
     padding: 1rem 1rem 0;
+    width: 100%;
   }
 
   .body {
